Fix authorization filter issues
- Remove `AuthenticationContextAuthorizationFilter` as it wasn't being called. Properly move functionality to `TgsAuthorizeAttribute`
- Fix unauth/forbid decision being skipped for tokens with password updates.
- Fix issues if a user used a token at the exact same second of a password update.
- Update auth documentation.
